pyproject.toml
[tool.poetry]
name = "csrbackend"
version = "0.1.0"
description = ""
authors = ["DragonPyli <codewizard001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.8"
Flask = "^2.0.1"
keras = "^2.4.3"
tensorflow = "^2.5.0"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.18.0"
pandas = "^1.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
